+++ backend/mora/service/manager.py
@@ -24,13 +24,16 @@ from .. import validator
 
 
 class ManagerRequestHandler(handlers.OrgFunkRequestHandler):
-    __slots__ = ()
+    __slots__ = ('termination_field', 'termination_value')
 
     role_type = 'manager'
     function_key = mapping.MANAGER_KEY
 
-    termination_field = mapping.USER_FIELD
-    termination_value = {}
+    def __init__(self, request, request_type):
+        self.termination_field = mapping.USER_FIELD
+        self.termination_value = {}
+
+        super().__init__(request, request_type)
 
     def prepare_create(self, req):
         """ To create a vacant manager postition, set employee_uuid to None
@@ -204,3 +207,13 @@ class ManagerRequestHandler(handlers.OrgFunkRequestHandler):
 
         self.payload = payload
         self.uuid = manager_uuid
+
+    def prepare_terminate(self, request: dict):
+        # If we want to terminate the managers as well
+        if request.get('request').get('terminate_all'):
+            self.termination_value = \
+                handlers.OrgFunkRequestHandler.termination_value
+            self.termination_field = \
+                handlers.OrgFunkRequestHandler.termination_field
+
+        super().prepare_terminate(request)