+++ frontend/src/components/MoTreeView/MoTreeView.vue
@@ -1,10 +1,13 @@
 <template>
   <div class="orgunit-tree">
     <liquor-tree
-      :ref="nameId"
-      :v-model="selected"
+      :ref="_nameId"
       :data="treeData"
-      :options="treeOptions">
+      :options="treeOptions"
+      @node:selected="onNodeSelected"
+      @node:checked="onNodeCheckedChanged"
+      @node:unchecked="onNodeCheckedChanged"
+    >
 
       <div class="tree-scope" slot-scope="{ node }">
         <template>
@@ -33,29 +36,64 @@ export default {
 
   props: {
     /**
-     * Defines a orgUuid.
+     * This control takes a string variable as its model, representing
+     * the UUID of the selected unit. Internally, the tree view does
+     * have access to reasonably full objects representing the unit,
+     * but they don't correspond _exactly_ to those used elsewhere, so
+     * we only pass the UUID.
+     *
+     * @model
      */
-    unitUuid: String,
+    value: {type: [String, Array]},
 
     /**
-     * Defines a atDate.
+     * Defines the date for rendering the tree; used for the time machine.
      */
-    atDate: [Date, String]
+    atDate: {type: [Date, String]},
+
+    /**
+     * UUID of unselectable unit.
+     */
+    disabledUnit: String,
+
+    /**
+     * Select more than one node
+     */
+    multiple: Boolean
   },
 
   computed: {
     ...mapGetters({
+      /**
+       * The tree view itself is dependent on the currently active
+       * organisation. Among other things, we should only show the units
+       * for that organisation, and also ensure that we reset the view
+       * whenever it changes.
+       */
       orgUuid: 'organisation/getUuid'
     }),
 
-    nameId () {
+    /**
+     * @private
+     */
+    _nameId () {
       return 'moTreeView' + this._uid
     },
 
+    /**
+     * Accessor for the LiquorTree instance.
+     *
+     * @protected
+     */
     tree () {
-      return this.$refs[this.nameId]
+      return this.$refs[this._nameId]
     },
 
+    /**
+     * A string representation of the currently rendered tree, useful
+     * for inspection and tests, with highlighting of expansion and
+     * selection states.
+     */
     contents () {
       function visitNode (node, level) {
         if (!node) {
@@ -66,13 +104,23 @@ export default {
             .map(c => visitNode(c, level))
         }
 
-        let text = node.selected() ? `=+= ${node.text} =+=` : node.text
+        let text = node.text
+
+        if (node.checked()) {
+          text = `\u2713 ${text}`
+        }
+
+        if (node.selected()) {
+          text = `=+= ${text} =+=`
+        }
+
+        if (node.disabled()) {
+          text = `~~~ ${text} ~~~`
+        }
 
         if (node.expanded()) {
           const r = {}
-
           r[text] = visitNode(node.children, level + 1)
-
           return r
         } else if (node.hasChildren()) {
           return '> ' + text
@@ -89,14 +137,27 @@ export default {
     let vm = this
 
     return {
+      /**
+       * LiquorTree model; required for some reason or other, but not
+       * actually used?
+       *
+       * @private
+       */
       treeData: [],
-      selected: undefined,
-      units: {},
 
+      /**
+       * LiquorTree options.
+       *
+       * @private
+       */
       treeOptions: {
-        minFetchDelay: 1,
         parentSelect: true,
-
+        multiple: false,
+        checkbox: this.multiple,
+        checkOnSelect: this.multiple,
+        autoDisableChildren: false,
+        autoCheckChildren: false,
+        minFetchDelay: 1,
         fetchData (node) {
           return vm.fetch(node)
         }
@@ -107,27 +168,7 @@ export default {
   mounted () {
     const vm = this
 
-    this.tree.$on('node:added', node => {
-      console.log(`TREE: adding node ${node.id}`)
-    })
-
-    this.tree.$on('node:removed', node => {
-      console.log(`TREE: removing node ${node.id}`)
-      delete vm.units[node.id]
-    })
-
-    this.tree.$on('node:selected', node => {
-      console.log(`TREE: selected node ${node.id}`)
-
-      vm.$emit('input', vm.units[node.id])
-    })
-
-    this.tree.$on('node:expanded', node => {
-      console.log('TREE: expanded', node.text, node.id)
-    })
-
     EventBus.$on('update-tree-view', () => {
-      console.log(`TREE: update tree view!`)
       vm.updateTree(true)
     })
 
@@ -135,34 +176,30 @@ export default {
   },
 
   watch: {
-    unitUuid (newVal, oldVal) {
-      console.log(`TREE: changing unit from ${oldVal} to ${newVal} (org=${this.orgUuid})`)
+    /**
+     * Update the selection when the value changes.
+     */
+    value (newVal, oldVal) {
+      if (JSON.stringify(newVal) === JSON.stringify(oldVal)) {
+        return
+      }
 
-      if (this.units && this.units[newVal]) {
-        this.setSelection(newVal)
-      } else if (newVal !== oldVal) {
+      const missing = this.tree.tree ? this.toArray(newVal).filter(
+        v => !this.tree.tree.getNodeById(v)
+      ) : []
+
+      if (missing.length) {
         this.updateTree()
+      } else {
+        this.setSelection(newVal)
       }
     },
 
-    selected: {
-      handler (newVal, oldVal) {
-        console.log(`TREE: selected changed to ${newVal}`)
-      },
-      deep: true
-    },
-
-    treeData: {
-      handler (newVal, oldVal) {
-        console.log(`TREE: treeData changed to ${newVal}`)
-      },
-      deep: true
-    },
-
+    /**
+     * Reset the selection when the organisation changes.
+     */
     orgUuid (newVal, oldVal) {
       let vm = this
-
-      console.log(`TREE: changing organisation from ${oldVal} to ${newVal}`)
 
       // in order to avoid updating twice, only do so when no unit
       // is configured; otherwise, we'll update when the unit clears
@@ -173,37 +210,110 @@ export default {
       //
       // yes, this is a bit of a hack :(
       setTimeout(() => {
-        if (oldVal || !vm.unitUuid) {
+        if (oldVal || !vm.value || vm.value.length === 0) {
           vm.updateTree(true)
         }
       }, 100)
     },
 
+    disabledUnit (newVal) {
+      for (const oldNode of this.tree.findAll({ disabled: true })) {
+        if (oldNode.id !== newVal) {
+          oldNode.enable()
+        }
+      }
+
+      let newNode = this.tree.tree.getNodeById(newVal)
+
+      if (newNode && newNode.enabled()) {
+        newNode.uncheck()
+        newNode.unselect()
+        newNode.disable()
+      }
+    },
+
+    /**
+     * Re-render the tree when the date changes.
+     */
     atDate () {
       this.updateTree()
     }
   },
 
   methods: {
-    /**
-     * Select the unit corresponding to the given ID, assuming it's present.
-     */
-    setSelection (unitid) {
-      if (!unitid) {
-        unitid = this.unitUuid
-      }
+    onNodeCheckedChanged (node) {
+      if (this.multiple) {
+        let checked = this.getSelection()
 
-      console.log(`TREE: selecting ${unitid}`)
-      this.tree.tree.unselectAll()
-
-      let n = this.tree.tree.getNodeById(unitid)
-
-      if (n) {
-        n.expandTop()
-        n.select()
+        this.$emit('input', checked)
       }
     },
 
+    /**
+     * Propagate the selection to the model.
+     *
+     * @protected
+     */
+    onNodeSelected (node) {
+      if (!this.multiple) {
+        this.$emit('input', node.id)
+      }
+    },
+
+    getSelection () {
+      let nodes = this.multiple ?
+          this.tree.checked() : this.tree.selected()
+      return this.toArray(nodes.map(n => n.id))
+    },
+
+    toArray(values) {
+      let vs = values ? values instanceof Array ? values : [values] : []
+
+      vs.sort()
+
+      return vs
+    },
+
+    /**
+     * Select the units corresponding to the given IDs, assuming
+     * they're present, and updating the tree otherwise.
+     */
+    setSelection (unitids) {
+      // wrap the values in a list, if necessary, handling absence
+      const newVal = this.toArray(unitids)
+      const oldVal = this.getSelection()
+
+      // handle removals
+      for (const uuid of oldVal.filter(v => !newVal.includes(v))) {
+        const node = this.tree.tree.getNodeById(uuid)
+
+        if (this.multiple) {
+          node.uncheck()
+        } else {
+          node.unselect()
+        }
+      }
+
+      // handle additions
+      for (const uuid of newVal.filter(v => !oldVal.includes(v))) {
+        const node = this.tree.tree.getNodeById(uuid)
+
+        if (node) {
+          node.expandTop()
+
+          if (this.multiple) {
+            node.check()
+          } else {
+            node.select()
+          }
+        }
+      }
+    },
+
+    /**
+     * Add the given node to the tree, nested under the parent, specified, or
+     * root otherwise.
+     */
     addNode (unit, parent) {
       let preexisting = this.tree.tree.getNodeById(unit.uuid)
 
@@ -222,6 +332,7 @@ export default {
         this.tree.append(this.toNode(unit))
       }
     },
+
     /**
      * Convert a unit object into a node suitable for adding to the
      * tree.
@@ -229,13 +340,15 @@ export default {
      * This method handles both eager and lazy loading of child nodes.
      */
     toNode (unit) {
-      this.units[unit.uuid] = unit
-
       return {
         text: unit.name,
         isBatch: unit.children ? false : unit.child_count > 0,
         id: unit.uuid,
-        children: unit.children ? unit.children.map(this.toNode.bind(this)) : null
+        children: unit.children ? unit.children.map(this.toNode.bind(this)) : null,
+        state: {
+          disabled: unit.uuid === this.disabledUnit,
+          expanded: Boolean(unit.children)
+        }
       }
     },
 
@@ -245,45 +358,36 @@ export default {
     updateTree (force) {
       let vm = this
 
-      if (!this.orgUuid || !this.tree) {
-        return
-      }
-
       if (force) {
         this.tree.remove({}, true)
-        this.units = {}
       }
 
-      if (this.unitUuid) {
-        OrganisationUnit.getAncestorTree(this.unitUuid, this.atDate)
+      if (this.multiple ? this.value.length > 0 : this.value) {
+        OrganisationUnit.getAncestorTree(this.value, this.atDate)
           .then(response => {
-            console.log('TREE: injecting unit tree', response.uuid)
-
             vm.addNode(response, null)
             vm.tree.sort()
-
-            vm.setSelection()
+            vm.setSelection(this.value)
           })
-      } else {
+      } else if (this.orgUuid) {
         Organisation.getChildren(this.orgUuid, this.atDate)
           .then(response => {
-            console.log('TREE: injecting org tree')
-
-            vm.units = {}
 
             for (let unit of response) {
               vm.addNode(unit, null)
             }
 
             vm.tree.sort()
+            vm.setSelection(this.value)
           })
       }
     },
 
+    /**
+     * LiquorTree lazy data fetcher.
+     */
     fetch (node) {
       let vm = this
-
-      console.log(`TREE: fetching ${node.text}`)
 
       if (!this.orgUuid || node.fetching) {
         // nothing to do, so return something that does nothing
@@ -298,13 +402,9 @@ export default {
       return OrganisationUnit.getChildren(node.id, this.atDate)
         .then(response => {
           node.fetching = false
-
           return response.map(vm.toNode.bind(vm))
         }).catch(error => {
-          console.error('fetch failed', error)
-
           node.fetching = false
-
           throw error
         })
     }
@@ -312,22 +412,24 @@ export default {
 }
 </script>
 
-<!-- this particular styling is not scoped, otherwise liqour tree cannot detect the overwrites -->
+<!-- this particular styling is not scoped, otherwise liqour tree
+     cannot detect the overwrites. to ensure that we _always_ win, we
+     increase the specificity of the selectors  -->
 <style>
-  .tree > .tree-root, .tree-content {
+  .orgunit-tree .tree > .tree-root, .tree-content {
      padding: 0;
    }
 
-   .tree-children {
+   .orgunit-tree .tree-children {
      transition-timing-function: ease-in-out;
      transition-duration: 150ms;
    }
 
-  .tree-node.selected > .tree-content {
-    background: #007bff;
+  .orgunit-tree .tree-node.selected > .tree-content {
+    background-color: #007bff;
   }
 
-  .tree-node.selected > .tree-content > .tree-anchor {
+  .orgunit-tree .tree-node.selected > .tree-content > .tree-anchor {
     color: #fff;
   }
 </style>