+++ frontend/src/employee/index.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="row">
     <div class="col-12 workflow-padding">
-      <router-view :key="$route.params.uuid"/>
+      <router-view :key="route.params.uuid"/>
 
       <mo-log/>
     </div>
@@ -14,7 +14,7 @@
 /**
  * A employee component.
  */
-
+import { mapState } from 'vuex'
 import MoEmployeeWorkflows from './MoEmployeeWorkflows/MoEmployeeWorkflows'
 import MoLog from '@/components/MoLog/MoLog'
 
@@ -22,6 +22,11 @@ export default {
   components: {
     MoEmployeeWorkflows,
     MoLog
+  },
+  computed: {
+    ...mapState({
+      route: 'route'
+    })
   }
 }
 </script>